+++ crates/jsonnet-analyze/src/util.rs
@@ -32,6 +32,7 @@ pub enum FormatEngine {
   /// - Is named `bin/jsonnetfmt` relative to the workspace root
   /// - Accepts a `-stdio` flag
   /// - Accepts a positional argument, the name of the file being formatted
+  /// - Has its working directory set to the workspace root
   /// - Takes in from stdin the unformatted file contents
   /// - Outputs to stdout the formatted file contents
   BinJsonnetFmtStdio,